+++ CHANGELOG.md
@@ -1,5 +1,25 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=MINOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* Add `TextPanel` class and builder.
+  text panel lets you make information and description panels etc. for your dashboards.
+  - `mode` - Here you can choose between markdown, HTML or text.
+  - `content` - Here you write your content.
+  ```kotlin
+  dashboard(title = "Dashboard with Text Panel") {
+      panels {
+          textPanel("Description") {
+              mode = ContentMode.MD
+              content = "### Text Panel with MD content"
+          }
+      }
+  }
+  ``` 
+* Update `Legend.kt` class:
+  - Added new field `sideWidth` Available when _To the right_ is checked. Value to control the minimum width for the legend (default 0)
+* Added `ConsolidateBy` graphite function and corresponding `consolidateBy()` extension functions for `Metric`
+  ```kotlin
+  val metric = "*.*.*.requests.incoming.*.*.process_time.*.count" consolidateBy MAX
+  ``` 
 ### NEXT_VERSION_DESCRIPTION_END
 ## [2.3.0]() (26-09-2019)